* versiyon numarası güncellendi, şu anki versiyon 0.6.7 * yeni user access control sistemi ufak eksiklerle kullanımda, artık panelde kontroller bu sistem ile yapılıyor, istenirse panel haricinde de custom roller tanımlanıp üyelik gerektiren işlemlerde kullanılabilir. access control işlemi için kullanılan fonksiyon ismi checkAccessStatus(). * aklıma gelmeyen ufak başka düzenlemeler yapıldı.

+++ system/classes/USER.php
@@ -1,15 +1,12 @@
 <?php
 class PA_USER extends PA_USER_TICKET
 {
-	private $table;
-	public $trackKeyName;
-	public $loggedInUser;
+	protected $table;
 	
 	function PA_USER()
 	{
 		parent::PA_USER_TICKET();
 		$this->table = $this->tables->user;
-		$this->loggedInUser = $this->getLoggedInUser();
 	}
 	
 	function completeRegistration($user_id, $username, $password)
@@ -22,9 +19,34 @@ class PA_USER extends PA_USER_TICKET
 		return $this->execute("UPDATE {$this->table} SET username=?, password=?, pass_key=?, status='active' WHERE user_id=?", array($username, $encryptedPassword, $pass_key, $user_id));
 	}
 	
-	function inviteAdminUser($displayname, $email, $user_type, $end_date = "0000-00-00 00:00:00")
+	function inviteUser($displayname, $email, $roles=null, $groups=null)
 	{
-		if($user_id = $this->insert($this->table, array("displayname"=>$displayname, "email"=>$email, "user_type"=>$user_type, "status"=>"invited")))
+		global $ADMIN;
+		
+		// Kullanıcıyı ekle
+		$user_id = $this->insert($this->table, array("displayname"=>$displayname, "email"=>$email, "status"=>"invited"));
+		
+		// Kullanıcının rollerini ekle
+		if(($roles != null) && is_array($roles))
+		{
+			$role_count = sizeof($roles);
+			for($i=0; $i<$role_count; $i++)
+			{
+				$ADMIN->USER_ROLE->addUserRole($user_id, $roles[$i]);
+			}
+		}
+		
+		// Kullanıcının guruplarını ekle
+		if(($groups != null) && is_array($groups))
+		{
+			$group_count = sizeof($groups);
+			for($i=0; $i<$group_count; $i++)
+			{
+				$ADMIN->USER_ROLE->addUserRole($user_id, $groups[$i]);
+			}
+		}
+		
+		if($user_id > 0)
 		{
 			return $this->sendInvitationMail($user_id);
 		}
@@ -32,15 +54,36 @@ class PA_USER extends PA_USER_TICKET
 			return false;
 	}
 	
+	function addUser($username, $displayname, $email, $password)
+	{
+		if($this->getUserCount() <= 0)
+		{
+				
+			global $secureKey;
+				
+			$pass_key = randomString(20);
+			$encryptedPassword = sha1($secureKey . $password . $pass_key);
+				
+			return $this->insert($this->table, array("username"=>$username, "displayname"=>$displayname, "password"=>$encryptedPassword, "pass_key"=>$pass_key, "email"=>$email, "register_time"=>"NOW()"));
+		}
+		else
+		{
+			$this->error = "Zaten en az bir kullanıcı mevcut, bu şekilde yeni bir kullanıcı oluşturamazsınız!";
+			return false;
+		}
+	}
+	
 	function sendInvitationMail($user_id, $end_date = "0000-00-00 00:00:00")
 	{
+		global $ADMIN;
+		
 		$user = $this->getUserById($user_id);
 		$this->closeTicketsByTicketType($user_id, "invitation");
 		$ticket_id = $this->openTicket($user_id, "invitation", $end_date);
 		$ticket = $this->selectTicket($ticket_id);
 		$site_title = get_option("admin_siteTitle");
 		$register_link = get_option("admin_siteAddress") . "/admin/complete_registration.php?type=invitation&user={$user_id}&key={$ticket->ticket_key}";
-		$invitation_sender = $this->loggedInUser; // Davetiyeyi gönderen kullanıcı
+		$invitation_sender = $ADMIN->AUTHENTICATION->authenticated_user; // Davetiyeyi gönderen kullanıcı
 			
 		$mesaj  = "Sayın  <b>{$user->displayname}</b>, <br /> ";
 		$mesaj .= "<b>{$invitation_sender->displayname}</b> kullanıcısı ";
@@ -57,47 +100,6 @@ class PA_USER extends PA_USER_TICKET
 	{
 		$user = $this->getUserByEmail($email);
 		return $this->sendInvitationMail($user->user_id);
-	}
-	
-	function login($username, $password, $captcha_used_correctly = false)
-	{
-		if($user = $this->getUserByUsername($username))
-		{
-			global $secureKey;
-			
-			$encryptedPassword = sha1($secureKey . $password . $user->pass_key);
-			
-			if($encryptedPassword == $user->password)
-			{
-				if(($user->captcha_limit > 0) || $captcha_used_correctly)
-				{
-					$this->resetUserCaptchaLimit($user->user_id);
-					$this->openTrack($user->user_id);
-					$this->loggedInUser = $user;
-					return true;
-				}
-				else
-				{
-					return "login_with_captcha";
-				}
-			}
-			else
-			{
-				$this->decreaseUserCaptchaLimit($user->user_id);
-				return false;
-			}
-		}
-		else
-			return false;
-	}
-	
-	function logout()
-	{
-		$tracking_key = $_SESSION[$this->trackKeyName];
-		$this->closeTrack($tracking_key);
-		unset($_SESSION[$this->trackKeyName]);
-		unset($this->loggedInUser);
-		header("Location:login.php");
 	}
 	
 	function changePassword($user_id, $password)
@@ -153,25 +155,6 @@ class PA_USER extends PA_USER_TICKET
 		}
 	}
 	
-	function createFirstAdminUser($username, $displayname, $email, $password)
-	{
-		if($this->getUserCount() <= 0)
-		{
-			
-			global $secureKey;
-			
-			$pass_key = randomString(20);
-			$encryptedPassword = sha1($secureKey . $password . $pass_key);
-			
-			return $this->insert($this->table, array("username"=>$username, "displayname"=>$displayname, "password"=>$encryptedPassword, "pass_key"=>$pass_key, "email"=>$email, "user_type"=>100, "register_time"=>"NOW()"));
-		}
-		else
-		{
-			$this->error = "Zaten en az bir kullanıcı mevcut, bu şekilde yeni bir kullanıcı oluşturamazsınız!";
-			return false;
-		}
-	}
-	
 	// TODO: username değerini değiştirme özelliği ekle
 	function updateUser($user_id, $image_id, $displayname, $birthday, $first_name, $last_name, $email, $phone, $password)
 	{	
@@ -192,50 +175,29 @@ class PA_USER extends PA_USER_TICKET
 		$variables[] = $user_id;
 		
 		if($this->execute($query, $variables))
-		{
-			$this->loggedInUser = $this->getLoggedInUser();
 			return true;
-		}
 		else
 			return false;
 	}
 	
-	function deleteUser($user_id, $delete_tracks = true)
+	// TODO: Silinen kullanıcı silinmeden önce sisteme giriş yapmışsa ve hala sistemdeyse onu sistemden de çıkarmanın yolunu bul.
+	function deleteUser($user_id)
 	{
-		if($delete_tracks && !$this->deleteTracksByUserId($user_id))
-			return false;
+		global $ADMIN;
 		
-		return $this->deleteUsersAllTickets($user_id) &&
-				$this->execute("DELETE FROM {$this->table} WHERE user_id=?", array($user_id));
+		return  $this->deleteUsersAllTickets($user_id) && $ADMIN->USER_ROLE->deleteUserRolesByUser($user_id) && $this->deleteTracksByUserId($user_id) &&
+				$ADMIN->USER_GROUP->deleteUserGroupsByUser($user_id) && $this->execute("DELETE FROM {$this->table} WHERE user_id=?", array($user_id));
 	}
 	
-	function deleteUserItself($user_id, $delete_tracks = true)
+	function deleteUserItself($user_id)
 	{
-		if($this->deleteUser($user_id, $delete_tracks))
+		if($this->deleteUser($user_id))
 		{
-			$this->logout();
+			global $ADMIN;
+			$ADMIN->AUTHENTICATION->logout();
 		}
 		else
 			return false;
-	}
-	
-	// TODO: burdaki işlemi tek sql sorgusu ile yapabilirsin.
-	function decreaseUserCaptchaLimit($user_id)
-	{
-		$captcha_limit = $this->get_value("SELECT captcha_limit FROM {$this->table} WHERE user_id=?", array($user_id));
-		$captcha_limit = intval($captcha_limit);
-
-		if($captcha_limit > 0)
-		{
-			return $this->execute("UPDATE {$this->table} SET captcha_limit=? WHERE user_id=?", array(($captcha_limit - 1), $user_id));
-		}
-		
-		return true;
-	}
-	
-	function resetUserCaptchaLimit($user_id)
-	{
-		return $this->execute("UPDATE {$this->table} SET captcha_limit=? WHERE user_id=?", array(3, $user_id));
 	}
 	
 	function getUserCount($status = "all")
@@ -284,14 +246,5 @@ class PA_USER extends PA_USER_TICKET
 		return $this->get_rows($query, $variables);
 	}
 	
-	private function getLoggedInUser()
-	{
-		$tracking_key = $_SESSION[$this->trackKeyName];
-		$track = $this->selectTrackByTrackingKey($tracking_key);
-		
-		if($track->status == "active")
-			return $this->getUserById($track->user_id);
-		else
-			return false;
-	}
+	
 }
